+++ spec/Bundle/DependencyInjection/Compiler/FosCacheResponseTaggerPassSpec.php
@@ -28,7 +28,7 @@ final class FosCacheResponseTaggerPassSpec extends ObjectBehavior
 
     public function let(): void
     {
-        $this->beConstructedWith('4.6.0@f264d7f0ce0b0d2ee58dbd1b9b9f87b9a3efa26b');
+        $this->beConstructedWith();
     }
 
     public function it_is_initializable(): void
@@ -48,21 +48,7 @@ final class FosCacheResponseTaggerPassSpec extends ObjectBehavior
         $this->process($container);
     }
 
-    public function it_doesnt_register_response_tagger_for_contao_before_4_6(ContainerBuilder $container): void
-    {
-        $this->beConstructedWith('4.5.9@f264d7f0ce0b0d2ee58dbd1b9b9f87b9a3efa26b');
-
-        $container->has(self::FOS_TAGGER_SERVICE_ID)
-            ->shouldBeCalled()
-            ->willReturn(true);
-
-        $container->setDefinition(self::TOOLKIT_TAGGER_SERVICE_ID, Argument::type(Definition::class))
-            ->shouldNotBeCalled();
-
-        $this->process($container);
-    }
-
-    public function it_registers_response_tagger_for_contao_4_6_if_fos_response_tagger_is_available(
+    public function it_registers_response_tagger_if_fos_response_tagger_is_available(
         ContainerBuilder $container
     ): void {
         $container->has(self::FOS_TAGGER_SERVICE_ID)