+++ src/Bundle/DependencyInjection/Compiler/FosCacheResponseTaggerPass.php
@@ -19,8 +19,6 @@ use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface as Comp
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Reference;
-use function explode;
-use function version_compare;
 
 /**
  * Class FosCacheResponseTaggerPass registers the FosCacheResponseTagger if it's supported.
@@ -32,32 +30,11 @@ final class FosCacheResponseTaggerPass implements CompilerPass
     private const SERVICE_ID = 'fos_http_cache.http.symfony_response_tagger';
 
     /**
-     * Contao core version in the format x.x.x.
-     *
-     * @var string
-     */
-    private $contaoCoreVersion;
-
-    /**
-     * FosCacheResponseTaggerPass constructor.
-     *
-     * @param string $contaoCoreVersion Contao core version in the format 1.0.0@hash.
-     */
-    public function __construct(string $contaoCoreVersion)
-    {
-        $this->contaoCoreVersion = explode('@', $contaoCoreVersion, 1)[0];
-    }
-
-    /**
      * {@inheritdoc}
      */
     public function process(ContainerBuilder $container): void
     {
         if (!$container->has(self::SERVICE_ID)) {
-            return;
-        }
-
-        if (version_compare($this->contaoCoreVersion, '4.6', '<')) {
             return;
         }